can now work with both dev and prod enviroments. User not authenticated error fixed.
