+++ mobile/app/register.tsx
@@ -17,7 +17,9 @@ export default function RegisterScreen() {
   const navigation = useNavigation<NavigationProp<RootStackParamList>>();
   const { useToast, useAlert } = showMessage();
   const API_BASE_URL =
-  Constants.expoConfig?.hostUri?.split(":").shift() ?? "localhost";
+  process.env.EXPO_PUBLIC_DEV_FLAG === "true"
+    ? `http://${Constants.expoConfig?.hostUri?.split(":").shift() ?? "localhost"}:8000`
+    : process.env.EXPO_PUBLIC_BACKEND_URL;
   
 
   const [firstName, setFirstName] = useState("");
@@ -45,7 +47,7 @@ export default function RegisterScreen() {
 
     try {
       
-      const response = await fetch(`${process.env.EXPO_PUBLIC_BACKEND_URL}/api/users/register/`, {
+      const response = await fetch(`${API_BASE_URL}/api/users/register/`, {
         method: "POST",
         headers: { "Content-Type": "application/json" },
         body: JSON.stringify({
